@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
-    <LangVersion>11</LangVersion>
+    <LangVersion>12</LangVersion>
     <OutputType>Exe</OutputType>
     <NoWarn>1573</NoWarn>
   </PropertyGroup>
