@@ -82,7 +82,7 @@ class HistoryRepository : IDisposable
     public List<HistoryEntry> GetWithStart(string start)
     {
         if (start.Length > 1000)
-            return new List<HistoryEntry>();
+            return [];
 
         var command = _db.CreateCommand();
         command.CommandText = """
@@ -100,7 +100,7 @@ class HistoryRepository : IDisposable
     public List<HistoryEntry> Search(string query)
     {
         if (query.Length > 1000)
-            return new List<HistoryEntry>();
+            return [];
 
         var command = _db.CreateCommand();
         command.CommandText = """
