@@ -30,7 +30,7 @@ class Repl
         var readLine = new ReadLinePrompt
         {
             HistoryHandler = historyHandler,
-            AutoCompletionHandler = new AutoCompleteHandler(shell, new[] { ' ', '/' }, highlightHandler),
+            AutoCompletionHandler = new AutoCompleteHandler(shell, [' ', '/'], highlightHandler),
             HighlightHandler = highlightHandler,
             HintHandler = new HintHandler(
                 historyRepository,
@@ -40,7 +40,7 @@ class Repl
             ),
             EnterHandler = new EnterHandler(),
             SearchHandler = new SearchHandler(historyRepository),
-            WordSeparators = new[] { ' ', '/', ':' },
+            WordSeparators = [' ', '/', ':'],
         };
 
         readLine.RegisterShortcut(
