@@ -2,10 +2,5 @@ using Elk.Parsing;
 
 namespace Elk.Interpreting.Exceptions;
 
-class RuntimeAccessLevelException : RuntimeException
-{
-    public RuntimeAccessLevelException(AccessLevel accessLevel, string symbolName)
-        : base($"Cannot access {accessLevel.ToString().ToLower()} symbol '{symbolName}'")
-    {
-    }
-}
+class RuntimeAccessLevelException(AccessLevel accessLevel, string symbolName)
+    : RuntimeException($"Cannot access {accessLevel.ToString().ToLower()} symbol '{symbolName}'");
