@@ -1,9 +1,3 @@
 namespace Elk.Interpreting.Exceptions;
 
-class RuntimeItemNotFoundException : RuntimeException
-{
-    public RuntimeItemNotFoundException(string item)
-        : base($"The item '{item}' was not found")
-    {
-    }
-}
+class RuntimeItemNotFoundException(string item) : RuntimeException($"The item '{item}' was not found");
