@@ -1,11 +1,3 @@
-using System.Collections.Generic;
-
 namespace Elk.Interpreting.Scope;
 
-class LocalScope : Scope
-{
-    public LocalScope(Scope? parent)
-        : base(parent)
-    {
-    }
-}
+class LocalScope(Scope? parent) : Scope(parent);
